feat(TDP-5056): add Dataset/DatasetMetadata conversion for TCOMPLocation
