+++ dataprep-dataset/src/main/java/org/talend/dataprep/dataset/adapter/conversion/DataSetMetadataBeanConversion.java
@@ -19,6 +19,7 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 import org.talend.dataprep.api.dataset.DataSetLocation;
 import org.talend.dataprep.api.dataset.DataSetMetadata;
+import org.talend.dataprep.api.dataset.location.LocalStoreLocation;
 import org.talend.dataprep.conversions.BeanConversionService;
 import org.talend.dataprep.dataset.adapter.Dataset;
 import org.talend.dataprep.processor.BeanConversionServiceWrapper;
@@ -29,11 +30,11 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import static org.talend.dataprep.conversions.BeanConversionService.fromBean;
 
 @Component
-public class DatasetBeanConversion extends BeanConversionServiceWrapper {
+public class DataSetMetadataBeanConversion extends BeanConversionServiceWrapper {
 
     private final ObjectMapper objectMapper;
 
-    public DatasetBeanConversion(ObjectMapper objectMapper) {
+    public DataSetMetadataBeanConversion(ObjectMapper objectMapper) {
         this.objectMapper = objectMapper;
     }
 
@@ -51,12 +52,17 @@ public class DatasetBeanConversion extends BeanConversionServiceWrapper {
                     dataset.setLabel(dataSetMetadata.getName());
 
                     DataSetLocation location = dataSetMetadata.getLocation();
-                    dataset.setType(location.getLocationType());
 
-                    ObjectNode jsonNode = objectMapper.createObjectNode();
-                    jsonNode.set("location", objectMapper.valueToTree(location));
-                    jsonNode.set("content", objectMapper.valueToTree(dataSetMetadata.getContent()));
-                    dataset.setProperties(jsonNode.toString());
+                    if (location instanceof LocalStoreLocation) {
+                        // FIXME by-pass for DataSet based on Local file (because Catalog doesn't provide Local file DataStore)
+                        dataset.setType(location.getLocationType());
+
+                        ObjectNode jsonNode = objectMapper.createObjectNode();
+                        jsonNode.set("location", objectMapper.valueToTree(location));
+                        jsonNode.set("content", objectMapper.valueToTree(dataSetMetadata.getContent()));
+
+                        dataset.setProperties(jsonNode.toString());
+                    }
 
                     return dataset;
                 }).build() //
